// interfaces/cart.ts
// Generated by https://quicktype.io

import { Decimal } from "@prisma/client/runtime";

export interface Cart {
    product: Product;
    user:    User;
}

export interface Product {
    id:        number;
    name:      string;
    image:     string;
    price:     Decimal;
    stock:     number;
}

export interface User {
    id:         number;
    email:      string;
    vendor:     string;
    username:   string;
}